account for lpstub section and add fillers
We add trvial code to the binary to ensure that the .text section is
larger than 2 MiB.

Additionally we modify the code responsible for finding the memory
range to re-map such that it subtracts the section containing the
code that does the actual remapping. Thus we can avoid throwing an
error if the section containing the remapping code is present inside
the segment we find.

To this end we rename the section containing the remapping code from
`.lpstub` to `lpstub`, because that will cause the linker to insert
symbols `__start_lpstub` and `__stop_lpstub` into the binary. We use
`__start_lpstub` to subtract the `lpstub` section, which contains the
code responsible for remapping.